[sim] Split RobotCore modified code between variants
