+++ RobotCore/src/sim/java/com/qualcomm/robotcore/wifi/NetworkConnectionFactory.java
@@ -1,0 +1,67 @@
+/*
+ * Copyright (c) 2016 Molly Nicholas
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without modification, are permitted
+ * (subject to the limitations in the disclaimer below) provided that the following conditions are
+ * met:
+ *
+ * Redistributions of source code must retain the above copyright notice, this list of conditions
+ * and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice, this list of conditions
+ * and the following disclaimer in the documentation and/or other materials provided with the
+ * distribution.
+ *
+ * Neither the name of Molly Nicholas nor the names of her contributors may be used to
+ * endorse or promote products derived from this software without specific prior written permission.
+ *
+ * NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS LICENSE. THIS
+ * SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
+ * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
+ * THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.qualcomm.robotcore.wifi;
+
+import android.content.Context;
+
+import com.qualcomm.robotcore.util.RobotLog;
+
+public class NetworkConnectionFactory {
+
+  public static final String NETWORK_CONNECTION_TYPE = "NETWORK_CONNECTION_TYPE";
+
+public static NetworkConnection getNetworkConnection(NetworkType type, Context context) {
+    RobotLog.v("Getting network assistant of type: " + type);
+    switch (type) {
+      case WIFIDIRECT:
+        return WifiDirectAssistant.getWifiDirectAssistant(context);
+      case LOOPBACK:
+        return null;  // not yet implemented
+      case SOFTAP:
+        return SoftApAssistant.getSoftApAssistant(context);
+      case WIRELESSAP:
+        return DriverStationAccessPointAssistant.getDriverStationAccessPointAssistant(context);
+      case RCWIRELESSAP:
+        return RobotControllerAccessPointAssistant.getRobotControllerAccessPointAssistant(context);
+      case EXTERNALAP:
+        return ExternalApAssistant.getExternalApAssistant(context);
+      case EMULATION_LOOPBACK:
+        return EmulationLoopbackAssistant.getEmulationLoopbackAssistant(context);
+      default:
+        return null;
+    }
+  }
+
+  public static NetworkType getTypeFromString(String type) {
+    return NetworkType.fromString(type);
+  }
+
+}
