+++ RobotCore/src/sim/java/com/qualcomm/robotcore/wifi/NetworkType.java
