+++ RobotCore/src/stock/java/com/qualcomm/robotcore/wifi/NetworkConnectionFactory.java
@@ -32,8 +32,6 @@ package com.qualcomm.robotcore.wifi;
 
 import android.content.Context;
 
-import com.dekaresearch.robotcore.simulation.EmulationLoopbackAssistant;
-import com.dekaresearch.robotcore.simulation.ExternalApAssistant;
 import com.qualcomm.robotcore.util.RobotLog;
 
 public class NetworkConnectionFactory {
@@ -53,10 +51,6 @@ public static NetworkConnection getNetworkConnection(NetworkType type, Context c
         return DriverStationAccessPointAssistant.getDriverStationAccessPointAssistant(context);
       case RCWIRELESSAP:
         return RobotControllerAccessPointAssistant.getRobotControllerAccessPointAssistant(context);
-      case EXTERNALAP:
-        return ExternalApAssistant.getExternalApAssistant(context);
-      case EMULATION_LOOPBACK:
-        return EmulationLoopbackAssistant.getEmulationLoopbackAssistant(context);
       default:
         return null;
     }
